docker/raster-uploader/requirements.txt
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
uWSGI==2.0.18
Werkzeug==2.0.2
GDAL==3.0.4
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.10.0
flake8==4.0.1
geoserver-rest==2.1.4
idna==3.3
kiwisolver==1.3.2
matplotlib==3.4.3
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
numpy==1.21.2
pandas==1.3.4
Pillow==8.4.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
python-decouple
